Moving notebooks around, other fixes
